+++ lib/basket_calculator.rb
@@ -19,12 +19,13 @@ module BasketCalculator
   end
 
 # going to need to change this method, so that it prints the prices and not a string, somehow adjust is so the post promotion price can be included. Convert it into a printing method.
-  def calculate_price(a_basket)
+  def calculate_price(a_basket, benefit)
     total = 0
     if a_basket.length == 0
       puts "Total price expected: £0.00"
     elsif a_basket.length == 1
       total += a_basket[0].price
+      total += benefit
       currency = number_to_currency(total)
       puts "Total price expected: #{currency}"
     else
@@ -32,6 +33,7 @@ module BasketCalculator
       a_basket.each { |x|
         total += x.price.to_i
       }
+      total += benefit
       currency = number_to_currency(total)
       puts "Total price expected: #{currency}"
     end
